+++ R/read.R
@@ -51,10 +51,12 @@
 #'
 #' # retrieve value labels
 #' mydat.val <- get_labels(mydat)}
-#'
 #' @importFrom haven read_sav read_sas read_dta
 #' @export
-read_spss <- function(path, atomic.to.fac = FALSE, drop.labels = FALSE, tag.na = FALSE, enc = NULL, verbose = TRUE) {
+read_spss <- function(path, atomic.to.fac = FALSE, drop.labels = FALSE, tag.na = FALSE, enc = NULL, verbose = FALSE) {
+  if (!requireNamespace("haven", quietly = TRUE)) {
+    stop("Package 'haven' required for this function. Please install it.")
+  }
   # read data file
   data.spss <- haven::read_sav(file = path, encoding = enc, user_na = tag.na)
   # prepare tagged NA?
@@ -155,7 +157,10 @@ read_spss <- function(path, atomic.to.fac = FALSE, drop.labels = FALSE, tag.na =
 
 #' @rdname read_spss
 #' @export
-read_sas <- function(path, path.cat = NULL, atomic.to.fac = FALSE, drop.labels = FALSE, enc = NULL, verbose = TRUE) {
+read_sas <- function(path, path.cat = NULL, atomic.to.fac = FALSE, drop.labels = FALSE, enc = NULL, verbose = FALSE) {
+  if (!requireNamespace("haven", quietly = TRUE)) {
+    stop("Package 'haven' required for this function. Please install it.")
+  }
   # read data file
   data <- haven::read_sas(data_file = path, catalog_file = path.cat, encoding = enc)
 
@@ -176,7 +181,10 @@ read_sas <- function(path, path.cat = NULL, atomic.to.fac = FALSE, drop.labels =
 
 #' @rdname read_spss
 #' @export
-read_stata <- function(path, atomic.to.fac = FALSE, drop.labels = FALSE, enc = NULL, verbose = TRUE) {
+read_stata <- function(path, atomic.to.fac = FALSE, drop.labels = FALSE, enc = NULL, verbose = FALSE) {
+  if (!requireNamespace("haven", quietly = TRUE)) {
+    stop("Package 'haven' required for this function. Please install it.")
+  }
   # read data file
   data <- haven::read_dta(file = path, encoding = enc)
 
@@ -218,7 +226,7 @@ read_stata <- function(path, atomic.to.fac = FALSE, drop.labels = FALSE, enc = N
   message("Converting atomic to factors. Please wait...\n")
   # iterate all columns
 
-  lapply(
+  as.data.frame(lapply(
     d,
     function(x) {
       # capture value labels attribute first
@@ -239,15 +247,14 @@ read_stata <- function(path, atomic.to.fac = FALSE, drop.labels = FALSE, enc = N
         if (!is.null(lab)) attr(x, "label") <- lab
       }
       x
-    }) %>%
-    as.data.frame(stringsAsFactors = FALSE)
+    }), stringsAsFactors = FALSE)
 }
 
 
 #' @importFrom tools file_ext
 #' @rdname read_spss
 #' @export
-read_data <- function(path, atomic.to.fac = FALSE, drop.labels = FALSE, enc = NULL, verbose = TRUE) {
+read_data <- function(path, atomic.to.fac = FALSE, drop.labels = FALSE, enc = NULL, verbose = FALSE) {
   switch(
     tools::file_ext(path),
     "sav" = ,