+++ R/set_na.R
@@ -64,6 +64,7 @@
 #' # add named vector as further missing value
 #' set_na(dummy, na = c("Refused" = 5), as.tag = TRUE)
 #' # see different missing types
+#' library(dplyr)
 #' library(haven)
 #' print_tagged_na(set_na(dummy, na = c("Refused" = 5), as.tag = TRUE))
 #'
@@ -143,7 +144,8 @@ set_na <- function(x, ..., na, drop.levels = TRUE, as.tag = FALSE) {
   }
 
   # evaluate arguments, generate data
-  .dat <- get_dot_data(x, rlang::quos(...))
+  dots <- as.character(match.call(expand.dots = FALSE)$`...`)
+  .dat <- .get_dot_data(x, dots)
 
   if (is.data.frame(x)) {
     # iterate variables of data frame
@@ -170,17 +172,17 @@ set_na <- function(x, ..., na, drop.levels = TRUE, as.tag = FALSE) {
 }
 
 
-#' @importFrom purrr map2
 #' @importFrom stats na.omit
-#' @importFrom haven tagged_na na_tag
 set_na_helper <- function(x, value, drop.levels, as.tag, var.name) {
   # check if values has only NA's
   if (sum(is.na(x)) == length(x)) return(x)
 
+  if (!requireNamespace("haven", quietly = TRUE)) {
+    stop("Package 'haven' required for this function. Please install it.")
+  }
+
   if (is.list(value)) {
-    lnames <- purrr::map2(value, names(value), ~ rep(.y, length(.x))) %>%
-      unlist() %>%
-      unname()
+    lnames <- unname(unlist(mapply(function(.x, .y) rep(.y, length(.x)), value, names(value), SIMPLIFY = FALSE)))
     value <- unlist(value)
     names(value) <- lnames
   }