+++ R/val_labels.R
@@ -1,12 +1,13 @@
-#' @importFrom rlang enexprs
-#' @importFrom purrr map_at
 #' @rdname set_labels
 #' @export
 val_labels <- function(x, ..., force.labels = FALSE, force.values = TRUE, drop.na = TRUE) {
+  if (!requireNamespace("rlang", quietly = TRUE)) {
+    stop("Package 'rlang' required for this function to work. Please install it.")
+  }
+
   # get dots
   .dots <- rlang::enexprs(...)
-  lang <- unlist(lapply(rlang::enexprs(...), is.language))
-  labels <- purrr::map_at(.dots, which(lang), eval)
+  labels <- lapply(.dots, function(i) if (is.language(i)) eval(i) else i)
 
   # select variables
   vars <- names(labels)